+++ ktlint-ruleset-standard/src/test/kotlin/com/pinterest/ktlint/ruleset/standard/ChainWrappingRuleTest.kt
@@ -2,6 +2,7 @@ package com.pinterest.ktlint.ruleset.standard
 
 import com.pinterest.ktlint.test.diffFileFormat
 import com.pinterest.ktlint.test.diffFileLint
+import com.pinterest.ktlint.test.format
 import com.pinterest.ktlint.test.lint
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
@@ -43,5 +44,39 @@ class ChainWrappingRuleTest {
                 """.trimIndent()
             )
         ).isEmpty()
+    }
+
+    // https://github.com/pinterest/ktlint/issues/1130
+    @Test
+    fun `format when conditions`() {
+        assertThat(
+            ChainWrappingRule().format(
+                """
+                fun test(foo: String?, bar: String?, baz: String?) {
+                    when {
+                        foo != null &&
+                            bar != null
+                            && baz != null -> {
+                        }
+                        else -> {
+                        }
+                    }
+                }
+                """.trimIndent()
+            )
+        ).isEqualTo(
+            """
+           fun test(foo: String?, bar: String?, baz: String?) {
+               when {
+                   foo != null &&
+                       bar != null &&
+                       baz != null -> {
+                   }
+                   else -> {
+                   }
+               }
+           }
+            """.trimIndent()
+        )
     }
 }